PORTALS-2435 - self-review changes
