+++ packages/synapse-react-client/src/components/SchemaDrivenAnnotationEditor/AnnotationEditorUtils.ts
@@ -1,13 +1,4 @@
-import {
-  ADDITIONAL_PROPERTY_FLAG,
-  PROPERTIES_KEY,
-  deepEquals,
-  FieldProps,
-  RJSFValidationError,
-  StrictRJSFSchema,
-  RJSFSchema,
-  FormContextType,
-} from '@rjsf/utils'
+import { RJSFValidationError } from '@rjsf/utils'
 import { flatMap, groupBy, isEmpty } from 'lodash-es'
 import { ENTITY_CONCRETE_TYPE } from '@sage-bionetworks/synapse-types'
 import { entityJsonKeys } from '../../utils/functions/EntityTypeUtils'
@@ -126,55 +117,5 @@ export function convertToArray<T>(value: T): Array<unknown> {
     return value.split(',').map(s => s.trim()) // split a string of comma-separated values, then trim whitespace
   } else {
     return [value]
-  }
-}
-
-/**
- * `componentDidUpdate` function for RJSF ObjectField.
- *
- * For an object, this will
- * - convert additionalProperties formData to arrays
- * - convert schema-defined formData from an array to a non-array if the schema type is not an array
- * @param props
- */
-export function objectFieldComponentDidUpdate<
-  T = any,
-  S extends StrictRJSFSchema = RJSFSchema,
-  F extends FormContextType = any,
->(props: FieldProps<T, S, F>) {
-  const { schema, formData, onChange } = props
-  const newFormData = { ...formData }
-  if (schema[PROPERTIES_KEY]) {
-    Object.entries(schema[PROPERTIES_KEY]).forEach(([key, propertySchema]) => {
-      const data = newFormData[key]
-      if (propertySchema[ADDITIONAL_PROPERTY_FLAG]) {
-        /**
-         * All additional properties should be converted to arrays.
-         *
-         * We need to convert it right away because the order of items is not stable, and seems to depend on if the item is an array or not
-         */
-        if (!Array.isArray(data)) {
-          newFormData[key] = convertToArray(data)
-        }
-      } else {
-        /**
-         * If the schema does not call for an array, but the formData is an array, then this will coerce it to a string.
-         *
-         * This can occur when a formData value is an additionalProperty, which we always treat as an array, then the key
-         * is added to the schema (e.g. conditionally).
-         */
-        if (
-          typeof propertySchema === 'object' &&
-          'type' in propertySchema &&
-          propertySchema.type !== 'array' &&
-          Array.isArray(data)
-        ) {
-          newFormData[key] = data.map(v => `${v}`).join(', ')
-        }
-      }
-    })
-    if (!deepEquals(formData, newFormData)) {
-      onChange(newFormData as T)
-    }
   }
 }
